+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add user can now save a picture
 - Add owner picture on shared list
 - Add invitation email on inexistent user
+- Add fingerprint authentication
 
 ### Changed